+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- DF-1284 Changed error message to indicate why DynamoDB schema overwrite doesn't work
 
 ## [0.14.0] - 2017-12-28
 ### Added