+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- DF-1275 Begin to add support for multi-column constraints
+- DF-1275 Initial support for multi-column constraints
 ### Changed
 - DF-1284 Changed error message to indicate why DynamoDB schema overwrite doesn't work
 ### Fixed