+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - DF-1287 Fixed NodeJS (and Python) script execution for large script. Made script size configurable
 - Fixed CORS
+- DF-1229 Added support for accented characters in api route
 
 ## [0.14.0] - 2017-12-28
 ### Added